update to MLA release (FILEIO v1.05) • Speed improvement when searching for free cluster to create new file. By starting at the last successfully found cluster rather than the start of the drive. • Fixes issue of writing after seeking causing trailing 0s in the file. • Fixes error in DirectoryRemove failing to correctly delete folder when folder contains previously deleted folders/files. • Fixes issue where a file marked with 8.3 designator not recognized as 8.3 file. • Speed improvements for the seek function. • Fixes issue where some drives may not mount successfully. • Fixes issue where format request may not happen successfully.

+++ inc/fileio.h
@@ -29,7 +29,7 @@ please contact mla_licensing@microchip.com
 #include "fileio_config.h"
 #include "system.h"
 
-#include <fileio_media.h>
+#include "fileio_media.h"
 
 
 /*******************************************************************/
@@ -404,9 +404,9 @@ typedef union
 {
     struct
     {
-        uint16_t day : 5;           // Day (1-31)
-        uint16_t month : 4;         // Month (1-12)
-        uint16_t year : 7;          // Year (number of years since 1980)
+        unsigned day : 5;           // Day (1-31)
+        unsigned month : 4;         // Month (1-12)
+        unsigned year : 7;          // Year (number of years since 1980)
     } bitfield;
     uint16_t value;
 } FILEIO_DATE;
@@ -416,9 +416,9 @@ typedef union
 {
     struct
     {
-        uint16_t secondsDiv2 : 5;   // (Seconds / 2) ( 1-30)
-        uint16_t minutes : 6;       // Minutes ( 1-60)
-        uint16_t hours : 5;         // Hours (1-24)
+        unsigned secondsDiv2 : 5;   // (Seconds / 2) ( 1-30)
+        unsigned minutes : 6;       // Minutes ( 1-60)
+        unsigned hours : 5;         // Hours (1-24)
     } bitfield;
     uint16_t value;
 } FILEIO_TIME;
